// Frontend/frontend/src/app/services/generated-api/conventions/model/ticketTemplateCreateDto.ts
/**
 * Conventions.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TicketTemplateCreateDto { 
    name?: string | null;
    description?: string | null;
    price?: number;
}